[FIX] 패키지 구조 변경 및 불필요한 클래스 제거
- 서비스의 복잡도 증가에 따라, 패키지구조를 도메인패키지 구조로 변경
- 보안/검증로직이 Gateway로 집중됨에 따라, 사용되지않는 PermissionType 제거
