--- src/main/kotlin/site/iplease/iadserver/advice/IpAssignDemandControllerAdvice.kt
@@ -1,25 +1,0 @@
-package site.iplease.iadserver.advice
-
-import org.springframework.http.ResponseEntity
-import org.springframework.web.bind.annotation.ExceptionHandler
-import org.springframework.web.bind.annotation.RestControllerAdvice
-import reactor.core.publisher.Mono
-import reactor.kotlin.core.publisher.toMono
-import site.iplease.iadserver.controller.IpAssignDemandController
-import site.iplease.iadserver.exception.WrongExpireDateException
-import site.iplease.iadserver.exception.WrongTitleException
-
-@RestControllerAdvice(basePackageClasses = [IpAssignDemandController::class])
-class IpAssignDemandControllerAdvice {
-    @ExceptionHandler(WrongExpireDateException::class)
-    fun handle(exception: WrongExpireDateException): Mono<ResponseEntity<String>> =
-        ResponseEntity.badRequest()
-            .body(exception.localizedMessage)
-            .toMono()
-
-    @ExceptionHandler(WrongTitleException::class)
-    fun handle(exception: WrongTitleException): Mono<ResponseEntity<String>> =
-        ResponseEntity.badRequest()
-            .body(exception.localizedMessage)
-            .toMono()
-}
