--- src/main/kotlin/site/iplease/iadserver/data/message/IpAssignDemandCreateMessage.kt
@@ -1,6 +1,0 @@
-package site.iplease.iadserver.data.message
-
-data class IpAssignDemandCreateMessage (
-    val demandId: Long,
-    val issuerId: Long
-)
