--- src/main/kotlin/site/iplease/iadserver/data/request/AssignIpDemandRequest.kt
@@ -1,11 +1,0 @@
-package site.iplease.iadserver.data.request
-
-import site.iplease.iadserver.data.type.AssignIpUsageType
-import java.time.LocalDate
-
-data class AssignIpDemandRequest (
-    val title: String,
-    val description: String,
-    val usage: AssignIpUsageType,
-    val expireAt: LocalDate,
-)
