--- src/main/kotlin/site/iplease/iadserver/data/response/AssignIpDemandResponse.kt
@@ -1,5 +1,0 @@
-package site.iplease.iadserver.data.response
-
-data class AssignIpDemandResponse (
-    val demandId: Long
-)
