--- src/main/kotlin/site/iplease/iadserver/data/type/AssignIpUsageType.kt
@@ -1,5 +1,0 @@
-package site.iplease.iadserver.data.type
-
-enum class AssignIpUsageType {
-    USE_NETWORK, DEPLOY_SERVER
-}
