--- src/main/kotlin/site/iplease/iadserver/data/type/MessageType.kt
@@ -1,7 +1,0 @@
-package site.iplease.iadserver.data.type
-
-enum class MessageType(
-    val routingKey: String
-) {
-    IP_ASSIGN_DEMAND_CREATE("ipAssignDemandCreate")
-}
