--- src/main/kotlin/site/iplease/iadserver/data/type/PermissionType.kt
@@ -1,5 +1,0 @@
-package site.iplease.iadserver.data.type
-
-enum class PermissionType {
-    GUEST, USER, OPERATOR, ADMINISTRATOR
-}
