+++ src/main/kotlin/site/iplease/iadserver/domain/demand/data/request/AssignIpDemandRequest.kt
@@ -1,0 +1,11 @@
+package site.iplease.iadserver.domain.demand.data.request
+
+import site.iplease.iadserver.domain.demand.data.type.AssignIpUsageType
+import java.time.LocalDate
+
+data class AssignIpDemandRequest (
+    val title: String,
+    val description: String,
+    val usage: AssignIpUsageType,
+    val expireAt: LocalDate,
+)
