+++ src/main/kotlin/site/iplease/iadserver/domain/demand/data/response/AssignIpDemandResponse.kt
@@ -1,0 +1,5 @@
+package site.iplease.iadserver.domain.demand.data.response
+
+data class AssignIpDemandResponse (
+    val demandId: Long
+)
