+++ src/main/kotlin/site/iplease/iadserver/domain/demand/data/type/AssignIpUsageType.kt
@@ -1,0 +1,5 @@
+package site.iplease.iadserver.domain.demand.data.type
+
+enum class AssignIpUsageType {
+    USE_NETWORK, DEPLOY_SERVER
+}
