+++ src/main/kotlin/site/iplease/iadserver/domain/demand/exception/WrongExpireDateException.kt
@@ -1,0 +1,5 @@
+package site.iplease.iadserver.domain.demand.exception
+
+import java.time.LocalDate
+
+class WrongExpireDateException(errorDetail: String, expireAt: LocalDate): RuntimeException("잘못된 만료일입니다! - $errorDetail - $expireAt")
