+++ src/main/kotlin/site/iplease/iadserver/domain/demand/exception/WrongTitleException.kt
@@ -1,0 +1,3 @@
+package site.iplease.iadserver.domain.demand.exception
+
+class WrongTitleException(detail: String): RuntimeException("잘못된 신청제목입니다! - $detail")
