+++ src/main/kotlin/site/iplease/iadserver/domain/demand/service/IpAssignDemandService.kt
@@ -1,0 +1,9 @@
+package site.iplease.iadserver.domain.demand.service
+
+import reactor.core.publisher.Mono
+import site.iplease.iadserver.domain.demand.data.dto.DemandDto
+
+interface IpAssignDemandService {
+    fun addDemand(demand: DemandDto): Mono<DemandDto>
+
+}
