--- src/main/kotlin/site/iplease/iadserver/exception/WrongTitleException.kt
@@ -1,3 +1,0 @@
-package site.iplease.iadserver.exception
-
-class WrongTitleException(detail: String): RuntimeException("잘못된 신청제목입니다! - $detail")
