+++ src/main/kotlin/site/iplease/iadserver/global/common/util/DateUtil.kt
@@ -1,0 +1,7 @@
+package site.iplease.iadserver.global.common.util
+
+import java.time.LocalDate
+
+interface DateUtil {
+    fun dateNow(): LocalDate
+}
