+++ src/main/kotlin/site/iplease/iadserver/global/common/util/StandardDateUtil.kt
@@ -1,0 +1,9 @@
+package site.iplease.iadserver.global.common.util
+
+import org.springframework.stereotype.Component
+import java.time.LocalDate
+
+@Component
+class StandardDateUtil: DateUtil {
+    override fun dateNow(): LocalDate = LocalDate.now()
+}
