+++ src/main/kotlin/site/iplease/iadserver/global/demand/data/message/IpAssignDemandCreateMessage.kt
@@ -1,0 +1,6 @@
+package site.iplease.iadserver.global.demand.data.message
+
+data class IpAssignDemandCreateMessage (
+    val demandId: Long,
+    val issuerId: Long
+)
