+++ src/main/kotlin/site/iplease/iadserver/infra/message/type/MessageType.kt
@@ -1,0 +1,7 @@
+package site.iplease.iadserver.infra.message.type
+
+enum class MessageType(
+    val routingKey: String
+) {
+    IP_ASSIGN_DEMAND_CREATE("ipAssignDemandCreate")
+}
