--- src/main/kotlin/site/iplease/iadserver/repository/DemandRepository.kt
@@ -1,6 +1,0 @@
-package site.iplease.iadserver.repository
-
-import org.springframework.data.repository.reactive.ReactiveCrudRepository
-import site.iplease.iadserver.data.entity.Demand
-
-interface DemandRepository: ReactiveCrudRepository<Demand, Long>
