--- src/main/kotlin/site/iplease/iadserver/service/IpAssignDemandService.kt
@@ -1,9 +1,0 @@
-package site.iplease.iadserver.service
-
-import reactor.core.publisher.Mono
-import site.iplease.iadserver.data.dto.DemandDto
-
-interface IpAssignDemandService {
-    fun addDemand(demand: DemandDto): Mono<DemandDto>
-
-}
