--- src/main/kotlin/site/iplease/iadserver/service/MessagePublishService.kt
@@ -1,8 +1,0 @@
-package site.iplease.iadserver.service
-
-import reactor.core.publisher.Mono
-import site.iplease.iadserver.data.type.MessageType
-
-interface MessagePublishService {
-    fun publish(type: MessageType, data: Any): Mono<Unit>
-}
