--- src/main/kotlin/site/iplease/iadserver/util/DateUtil.kt
@@ -1,7 +1,0 @@
-package site.iplease.iadserver.util
-
-import java.time.LocalDate
-
-interface DateUtil {
-    fun dateNow(): LocalDate
-}
