--- src/main/kotlin/site/iplease/iadserver/util/DemandDataValidator.kt
@@ -1,8 +1,0 @@
-package site.iplease.iadserver.util
-
-import reactor.core.publisher.Mono
-import site.iplease.iadserver.data.dto.DemandDto
-
-interface DemandDataValidator {
-    fun validate(dto: DemandDto): Mono<DemandDto>
-}
