--- src/main/kotlin/site/iplease/iadserver/util/StandardDateUtil.kt
@@ -1,9 +1,0 @@
-package site.iplease.iadserver.util
-
-import org.springframework.stereotype.Component
-import java.time.LocalDate
-
-@Component
-class StandardDateUtil: DateUtil {
-    override fun dateNow(): LocalDate = LocalDate.now()
-}
